+++ src/ResourceManager/DataMigration/Commands.DataMigration/Cmdlets/ValidateSqlServerSqlDbMiTaskCmdlet.cs
@@ -24,6 +24,7 @@ namespace Microsoft.Azure.Commands.DataMigration.Cmdlets
         private readonly string SelectedDatabase = "SelectedDatabase";
         private readonly string BackupBlobSasUri = "BackupBlobSasUri";
         private readonly string BackupFileShare = "BackupFileShare";
+        private readonly string BackupMode = "BackupMode";
 
         public ValidateSqlServerSqlDbMiTaskCmdlet(InvocationInfo myInvocation) : base(myInvocation)
         {
@@ -32,17 +33,21 @@ namespace Microsoft.Azure.Commands.DataMigration.Cmdlets
         public override void CustomInit()
         {
             this.TargetConnectionInfoParam(true);
+            this.SourceConnectionInfoParam(true);
             this.SimpleParam(SelectedDatabase, typeof(MigrateSqlServerSqlMIDatabaseInput[]), "Selected database to migrate", true);
             this.SimpleParam(BackupFileShare, typeof(FileShare), "File Share where the source server database files should be backed up. Use fully qualified domain name for the server", false);
             this.SimpleParam(BackupBlobSasUri, typeof(string), "SAS URI that provides DMS access to your storage account container that DMS will upload the backup files to and use for migrating databases to SQL DB Managed instance", true);
+            this.SimpleParam(BackupMode, typeof(string), "Backup Mode to specify whether to use existing backup or create new backup. Possible values include: 'CreateBackup', 'ExistingBackup'", false);
         }
 
         public override ProjectTaskProperties ProcessTaskCmdlet()
         {
             SqlConnectionInfo targetConnectionInfo = null;
+            SqlConnectionInfo sourceConnectionInfo = null;
             List<MigrateSqlServerSqlMIDatabaseInput> selectedDatabases = null;
             BlobShare backupBlobShare = null;
             FileShare backupFileShare = null;
+            string backupMode = "CreateBackup";
 
             if (MyInvocation.BoundParameters.ContainsKey(TargetConnection))
             {
@@ -50,6 +55,14 @@ namespace Microsoft.Azure.Commands.DataMigration.Cmdlets
                 PSCredential cred = (PSCredential)MyInvocation.BoundParameters[TargetCred];
                 targetConnectionInfo.UserName = cred.UserName;
                 targetConnectionInfo.Password = Decrypt(cred.Password);
+            }
+
+            if (MyInvocation.BoundParameters.ContainsKey(SourceConnection))
+            {
+                sourceConnectionInfo = (SqlConnectionInfo)MyInvocation.BoundParameters[SourceConnection];
+                PSCredential cred = (PSCredential)MyInvocation.BoundParameters[SourceCred];
+                sourceConnectionInfo.UserName = cred.UserName;
+                sourceConnectionInfo.Password = Decrypt(cred.Password);
             }
 
             if (MyInvocation.BoundParameters.ContainsKey(SelectedDatabase))
@@ -71,14 +84,21 @@ namespace Microsoft.Azure.Commands.DataMigration.Cmdlets
                 backupFileShare = (FileShare)MyInvocation.BoundParameters[BackupFileShare];
             }
 
+            if (MyInvocation.BoundParameters.ContainsKey(BackupMode))
+            {
+                backupMode = MyInvocation.BoundParameters[BackupMode] as string;
+            }
+
             var properties = new ValidateMigrationInputSqlServerSqlMITaskProperties
             {
                 Input = new ValidateMigrationInputSqlServerSqlMITaskInput
                 {
                     TargetConnectionInfo = targetConnectionInfo,
+                    SourceConnectionInfo = sourceConnectionInfo,
                     SelectedDatabases = selectedDatabases,
                     BackupBlobShare = backupBlobShare,
-                    BackupFileShare = backupFileShare
+                    BackupFileShare = backupFileShare,
+                    BackupMode = backupMode
                 }
             };