+++ src/StackAdmin/Azs.Subscriptions.Admin/Module/Azs.Subscriptions.Admin/Generated.PowerShell.Commands/SwaggerDefinitionCommands/New-AddonPlanDefinitionObject.ps1
@@ -16,19 +16,25 @@ Licensed under the MIT License. See License.txt in the project root for license 
 .PARAMETER MaxAcquisitionCount
     Maximum number of instances that can be acquired by a single subscription. If not specified, the assumed value is 1.
 
+.EXAMPLE
+
+    New-AddonPlanDefinitionObject -PlanId $planIdentifier -MaxAcquisitionCount 500
+
+    Create a new plan definition object for the specified plan with the acquisition limit of 500.
+
 #>
 function New-AddonPlanDefinitionObject
 {
-    param(    
+    param(
         [Parameter(Mandatory = $false)]
         [string]
         $PlanId,
-    
+
         [Parameter(Mandatory = $false)]
         [System.Nullable`1[long]]
         $MaxAcquisitionCount
     )
-    
+
     $Object = New-Object -TypeName Microsoft.AzureStack.Management.Subscriptions.Admin.Models.AddonPlanDefinition -ArgumentList @($planId,$maxAcquisitionCount)
 
     if(Get-Member -InputObject $Object -Name Validate -MemberType Method)